Add "Note" line into markdown output
Linkfix in syncLang.py

Issue #3172 Added check for existing files when renaming linked file

Added a check in LinkedFileViewModel.rename() to see if a file which is
being renamed already exists in the target location.  If such a file is
found then an error popup occurs informing the user of the conflict.

Ran localization update.

updated changelog